requirements.txt
google-generativeai==0.3.0
flask==2.2.3
python-dotenv==1.0.0
requests==2.28.2
murf==1.0.0
